Рефакторинг: Перемещение моделей RegisterModel и LoginModel в отдельные файлы в директории Models
