+++ UniversityStats/UniversityStats.API/Controllers/AuthController.cs
@@ -5,6 +5,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 using UniversityStats.Infrastructure.Identity;
+using UniversityStats.API.Models;
 
 namespace UniversityStats.API.Controllers;
 
@@ -94,18 +95,4 @@ public class AuthController : ControllerBase
 
         return new JwtSecurityTokenHandler().WriteToken(token);
     }
-}
-
-public class RegisterModel
-{
-    public required string Email { get; set; }
-    public required string Password { get; set; }
-    public required string FirstName { get; set; }
-    public required string LastName { get; set; }
-}
-
-public class LoginModel
-{
-    public required string Email { get; set; }
-    public required string Password { get; set; }
 }
