+++ UniversityStats/UniversityStats.API/Models/LoginModel.cs
@@ -1,0 +1,8 @@
+namespace UniversityStats.API.Models
+{
+    public class LoginModel
+    {
+        public required string Email { get; set; }
+        public required string Password { get; set; }
+    }
+}
