+++ UniversityStats/UniversityStats.API/Models/RegisterModel.cs
@@ -1,0 +1,10 @@
+namespace UniversityStats.API.Models
+{
+    public class RegisterModel
+    {
+        public required string Email { get; set; }
+        public required string Password { get; set; }
+        public required string FirstName { get; set; }
+        public required string LastName { get; set; }
+    }
+}
